+++ test/iterator.js
@@ -918,6 +918,26 @@ metatests.testSync('Iterator.indices on object with length', test => {
   test.strictSame(actual, [0, 1, 2, 3]);
 });
 
+metatests.testSync('Iterator.last empty', test => {
+  const actual = iter([]).last();
+  test.strictSame(actual, undefined);
+});
+
+metatests.testSync('Iterator.last numbers', test => {
+  const actual = iter([1, 2, 3, 4]).last();
+  test.strictSame(actual, 4);
+});
+
+metatests.testSync('Iterator.last objects', test => {
+  const actual = iter([{ a: 1 }, { a: 2 }, { a: 42 }]).last();
+  test.strictSame(actual, { a: 42 });
+});
+
+metatests.testSync('Iterator.last with default', test => {
+  const actual = iter([]).last(42);
+  test.strictSame(actual, 42);
+});
+
 metatests.testSync('iterEntries must iterate over object entries', test => {
   const source = { a: 13, b: 42, c: 'hello' };
   test.strictSame(iterEntries(source).toArray(), Object.entries(source));