+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased][unreleased]
 
+## [3.0.1][] - 2021-01-18
+
+- Optimize buffering crypto random generator
+
 ## [3.0.0][] - 2021-01-06
 
 - Forked from https://github.com/metarhia/common after 2.2.0
@@ -12,5 +16,7 @@
 See this link for 2.2.0 and all previous versions:
 https://github.com/metarhia/common/blob/master/CHANGELOG.md
 
-[unreleased]: https://github.com/metarhia/metautil/compare/v2.2.0...HEAD
+[unreleased]: https://github.com/metarhia/metautil/compare/v3.0.1...HEAD
+[3.0.1]: https://github.com/metarhia/metautil/compare/v3.0.0...v3.0.1
+[3.0.0]: https://github.com/metarhia/metautil/compare/v2.2.0...v3.0.0
 [2.2.0]: https://github.com/metarhia/common/releases/tag/v2.2.0